Fix MinGW compilation errors in jerry-ext (#4510)
MinGW also needs linking to ws2_32.
Defining socket types for unix/win32 for compatibility.
Also fixes #4512

JerryScript-DCO-1.0-Signed-off-by: Yonggang Luo luoyonggang@gmail.com

+++ jerry-ext/handle-scope/handle-scope-allocator.c
@@ -126,7 +126,7 @@ jerryx_handle_scope_get_child (jerryx_handle_scope_t *scope)
   {
     return (jerryx_handle_scope_t *) jerryx_handle_scope_pool.start;
   }
-  long idx = JERRYX_HANDLE_SCOPE_PRELIST_IDX (scope);
+  ptrdiff_t idx = JERRYX_HANDLE_SCOPE_PRELIST_IDX (scope);
   if (idx < 0)
   {
     return NULL;
